FE-1552: Share one surface shell and sampling walk between the sweep and optimization surfaces
