+++ libs/@hashintel/petrinaut/src/react/index.ts
@@ -67,6 +67,7 @@ export type {
   ExperimentRecord,
   ExperimentsActionsValue,
   ExperimentStatus,
+  ExperimentsActionsValue,
   ExperimentsContextValue,
 } from "./experiments/context";
 export { NotificationsContext } from "./notifications/context";
